+++ changelog.md
@@ -2,6 +2,7 @@
 1. Opt: Add ImmutableMultiDict, Auth: Request -> HttpConnection, graphql: executor_class
 2. database support: mysql
 3. applications: auto scan plugins
+4. applications: support subdomain router
 
 # v0.2.4
 1. support template: default jinja2